refs #154 missing state transition, 64-bit compilation issue
git-svn-id: https://devel.codership.com/svn/galera/trunk@1797 9148e567-b586-4176-b9b8-4e36cb4ec61b

+++ galera/src/gcs.hpp
@@ -19,7 +19,7 @@ namespace galera
     public:
         GcsI() { }
         virtual ~GcsI() { }
-        virtual int connect(const std::string& cluster_name,
+        virtual ssize_t connect(const std::string& cluster_name,
                             const std::string& cluster_url) = 0;
         virtual ssize_t set_initial_position(const wsrep_uuid_t& uuid,
                                              gcs_seqno_t seqno) = 0;
@@ -118,7 +118,7 @@ namespace galera
             last_applied_(GCS_SEQNO_ILL)
         { }
         ~DummyGcs() {}
-        int connect(const std::string& cluster_name,
+        ssize_t connect(const std::string& cluster_name,
                     const std::string& cluster_url) { return 0; }
         ssize_t set_initial_position(const wsrep_uuid_t& uuid,
                                      gcs_seqno_t seqno) { return 0; }
